Moved Steer/RunContext to SimulationDataFormat
Detectors and MC-analysis code should have an access to RunContext. Keeping RunContext in the
Steer creates circular dependencies, since the latter in turn depends on the detectors code
