+++ DataFormats/simulation/CMakeLists.txt
@@ -16,6 +16,7 @@ Set(HEADERS
     include/${MODULE_NAME}/MCCompLabel.h
     include/${MODULE_NAME}/TrackReference.h
     include/${MODULE_NAME}/PrimaryChunk.h
+    include/${MODULE_NAME}/RunContext.h
 )
 
 Set(LINKDEF src/SimulationDataLinkDef.h)
