+++ Steer/include/Steer/HitProcessingManager.h
@@ -11,7 +11,9 @@
 #ifndef O2_HITPROCESSINGMANAGER_H
 #define O2_HITPROCESSINGMANAGER_H
 
-#include <Steer/InteractionSampler.h>
+#include "CommonDataFormat/InteractionRecord.h"
+#include "SimulationDataFormat/RunContext.h"
+#include "Steer/InteractionSampler.h"
 #include <TGeoManager.h>
 #include <string>
 #include <vector>
@@ -25,61 +27,7 @@ namespace o2
 namespace steer
 {
 
-// a structure describing EventPart
-// (an elementary constituent of a collision)
-struct EventPart {
-  EventPart() = default;
-  EventPart(int s, int e) : sourceID(s), entryID(e) {}
-  int sourceID = 0; // the ID of the source (0->backGround; > 1 signal source)
-  // the sourceID should correspond to the chain ID
-  int entryID = 0; // the event/entry ID inside the chain corresponding to sourceID
-
-  static bool isSignal(EventPart e) { return e.sourceID > 1; }
-  static bool isBackGround(EventPart e) { return !isSignal(e); }
-  ClassDefNV(EventPart, 1);
-};
-
-// class fully describing the Collision contexts
-class RunContext
-{
- public:
-  RunContext() : mNofEntries{ 0 }, mMaxPartNumber{ 0 }, mEventRecords(), mEventParts() {}
-
-  TBranch* getBranch(std::string_view name, int sourceid = 0) const
-  {
-    if (mChains[sourceid]) {
-      return mChains[sourceid]->GetBranch(name.data());
-    }
-    return nullptr;
-  }
-
-  int getNCollisions() const { return mNofEntries; }
-
-  void setMaxNumberParts(int maxp) { mMaxPartNumber = maxp; }
-  int getMaxNumberParts() const { return mMaxPartNumber; }
-
-  const std::vector<o2::InteractionRecord>& getEventRecords() const { return mEventRecords; }
-  const std::vector<std::vector<EventPart>>& getEventParts() const { return mEventParts; }
-  const std::vector<TChain*>& getChains() const { return mChains; }
-
-  void printCollisionSummary() const;
-
- private:
-  int mNofEntries = 0;
-  int mMaxPartNumber = 0; // max number of parts in any given collision
-  std::vector<o2::InteractionRecord> mEventRecords;
-  // for each collision we record the constituents (which shall not exceed mMaxPartNumber)
-  std::vector<std::vector<EventPart>> mEventParts;
-  std::vector<TChain*> mChains; //! pointers to input chains
-
-  // it would also be appropriate to record the filenames
-  // that went into the chain
-
-  friend class HitProcessingManager;
-  ClassDefNV(RunContext, 1);
-};
-
-using RunFunct_t = std::function<void(const RunContext&)>;
+using RunFunct_t = std::function<void(const o2::steer::RunContext&)>;
 
 /// O2 specific run class; steering hit processing
 class HitProcessingManager
@@ -113,7 +61,7 @@ class HitProcessingManager
   // if -1 and only background chain will do number of entries in chain
   void setupRun(int ncollisions = -1);
 
-  const RunContext& getRunContext() { return mRunContext; }
+  const o2::steer::RunContext& getRunContext() { return mRunContext; }
 
   // serializes the runcontext to file
   void writeRunContext(const char* filename) const;
@@ -125,7 +73,7 @@ class HitProcessingManager
   bool setupChain();
 
   std::vector<RunFunct_t> mRegisteredRunFunctions;
-  RunContext mRunContext;
+  o2::steer::RunContext mRunContext;
 
   // this should go into the RunContext --> the manager only fills it
   std::vector<std::string> mBackgroundFileNames;
@@ -142,8 +90,8 @@ class HitProcessingManager
 
 inline void HitProcessingManager::sampleCollisionTimes()
 {
-  mRunContext.mEventRecords.resize(mRunContext.mNofEntries);
-  mInteractionSampler.generateCollisionTimes(mRunContext.mEventRecords);
+  mRunContext.getEventRecords().resize(mRunContext.getNCollisions());
+  mInteractionSampler.generateCollisionTimes(mRunContext.getEventRecords());
 }
 
 inline void HitProcessingManager::sampleCollisionConstituents()
@@ -177,10 +125,10 @@ inline void HitProcessingManager::sampleCollisionConstituents()
   };
 
   // we fill mRunContext.mEventParts
-  auto& eventparts = mRunContext.mEventParts;
+  auto& eventparts = mRunContext.getEventParts();
   eventparts.clear();
-  eventparts.resize(mRunContext.mEventRecords.size());
-  for (int i = 0; i < mRunContext.mEventRecords.size(); ++i) {
+  eventparts.resize(mRunContext.getEventRecords().size());
+  for (int i = 0; i < mRunContext.getEventRecords().size(); ++i) {
     eventparts[i].clear();
     // push any number of constituents?
     // for the moment just 2 : one background and one signal
