+++ Steer/src/SteerLinkDef.h
@@ -15,10 +15,6 @@
 #pragma link off all functions;
 
 #pragma link C++ class o2::steer::InteractionSampler+;
-#pragma link C++ class o2::steer::RunContext+;
-#pragma link C++ class o2::steer::EventPart+;
-#pragma link C++ class vector<o2::steer::EventPart>+;
-#pragma link C++ class vector<vector<o2::steer::EventPart>>+;
 #pragma link C++ class o2::steer::HitProcessingManager;
 #pragma link C++ class o2::steer::O2RunSim+;
 #pragma link C++ class o2::steer::O2MCApplication+;
